refactor(server): extract audio streaming into server_stream.go
Part of #61 M1. Moves streamAudio, generateAndSendChunk,
addClientToStream, notifyStreamEnd, negotiateCodec, and stateless
encoding helpers into a dedicated file.

These stay as Server methods because they own playback orchestration,
not role business logic. Mechanical move, no behavior change.
